Moved file with utility functions
